funding_arbitrage_bot/requirements.txt
httpx>=0.24.0
websockets>=11.0.0
asyncio>=3.4.3
pyyaml>=6.0
hyperliquid-python-sdk>=0.0.4
rich>=13.0.0 